refactor to get the quiz-results on its own quiz-results page
