+++ pages/quiz-page/index.js
@@ -2,10 +2,12 @@ import useSWR from "swr";
 import { useState } from "react";
 import QuizForm from "@/components/QuizForm";
 import IntroText from "@/components/IntroText";
+import { useRouter } from "next/router";
 
 const fetcher = (url) => fetch(url).then((response) => response.json());
 
-export default function QuizPage() {
+export default function QuizPage({ assignedPoints, handleAssignPointsToDog }) {
+  const router = useRouter();
   // State (memory) for the answer values of the three questions
   const [answerValues, setAnswerValues] = useState({});
   const userValues = new URLSearchParams({
@@ -13,49 +15,49 @@ export default function QuizPage() {
     trainability: answerValues.trainability,
     energy: answerValues.energy,
   });
-
+  console.log("data", assignedPoints);
   // Fetch Data for all questions (answers)
   const { data } = useSWR(`/api/quizResults?${userValues}`, fetcher);
 
   // This state array is empty at the beginning and will be filled piece by piece
-  const [assignedPoints, setAssignedPoints] = useState([]);
+  // const [assignedPoints, setAssignedPoints] = useState([]);
 
   // Solution 1 (derived state)
-  const filteredAssignedPoints = assignedPoints;
-  filteredAssignedPoints.sort((a, b) => b.points - a.points);
+  // const filteredAssignedPoints = assignedPoints;
+  // filteredAssignedPoints.sort((a, b) => b.points - a.points);
 
-  function assignPointsToDog(dog) {
-    setAssignedPoints((assignedPoints) => {
-      const foundDog = assignedPoints.find(
-        (dogItem) => dogItem.name === dog.name
-      );
-      if (foundDog) {
-        return assignedPoints.map((dogItem) =>
-          dogItem.name === dog.name
-            ? { ...dogItem, points: dogItem.points + 1 }
-            : dogItem
-        );
-      }
-      // if the dog did not exist before
-      return [...assignedPoints, { name: dog.name, points: 1 }];
-    });
-  }
+  // function assignPointsToDog(dog) {
+  //   setAssignedPoints((assignedPoints) => {
+  //     const foundDog = assignedPoints.find(
+  //       (dogItem) => dogItem.name === dog.name
+  //     );
+  //     if (foundDog) {
+  //       return assignedPoints.map((dogItem) =>
+  //         dogItem.name === dog.name
+  //           ? { ...dogItem, points: dogItem.points + 1 }
+  //           : dogItem
+  //       );
+  //     }
+  //     // if the dog did not exist before
+  //     return [...assignedPoints, { name: dog.name, points: 1 }];
+  //   });
+  // }
 
   // Algorithmus
   function calculateAssignedPoints() {
     data[0].forEach((dog) => {
       // question #1 (barking)
-      assignPointsToDog(dog);
+      handleAssignPointsToDog(dog);
       // setAssignedPoints for these dogs +1
     });
     data[1].forEach((dog) => {
       // question #2 (energy)
-      assignPointsToDog(dog);
+      handleAssignPointsToDog(dog);
       // setAssignedPoints for these dogs +1
     });
     data[2].forEach((dog) => {
       // question #3 (trainability)
-      assignPointsToDog(dog);
+      handleAssignPointsToDog(dog);
       // setAssignedPoints for these dogs +1
     });
   }
@@ -68,13 +70,14 @@ export default function QuizPage() {
     setAnswerValues({ ...data });
     // Assign Points to the corresponding dog
     calculateAssignedPoints();
+    router.push("/quiz-results");
   }
 
   return (
     <div>
       <IntroText />
       <QuizForm onSubmit={handleSubmit} />
-      <h2>List of dogs with their points</h2>
+      {/* <h2>List of dogs with their points</h2>
       <ul>
         {filteredAssignedPoints.map((dog) => {
           return (
@@ -84,7 +87,7 @@ export default function QuizPage() {
           );
         })}
       </ul>
-      <p>Total dogs: {assignedPoints.length}</p>
+      <p>Total dogs: {assignedPoints.length}</p> */}
     </div>
   );
 }
